Split cvd/instances/instance_database_utils into config_path and data_viewer
"utils" is not descriptive, and the methods in these files fell into two
categories.

Bug: b/419645813
