+++ base/cvd/cuttlefish/host/commands/cvd/cli/commands/BUILD.bazel
@@ -462,7 +462,7 @@ cc_library(
         "//cuttlefish/host/commands/cvd/cli/commands:command_handler",
         "//cuttlefish/host/commands/cvd/cli/selector",
         "//cuttlefish/host/commands/cvd/instances",
-        "//cuttlefish/host/commands/cvd/instances:instance_database_utils",
+        "//cuttlefish/host/commands/cvd/instances:config_path",
         "//libbase",
     ],
 )
