+++ base/cvd/cuttlefish/host/commands/cvd/cli/selector/selector_common_parser.cpp
@@ -27,7 +27,7 @@
 #include "cuttlefish/common/libs/utils/contains.h"
 #include "cuttlefish/common/libs/utils/flag_parser.h"
 #include "cuttlefish/host/commands/cvd/cli/selector/selector_constants.h"
-#include "cuttlefish/host/commands/cvd/instances/instance_database_utils.h"
+#include "cuttlefish/host/commands/cvd/instances/device_name.h"
 
 namespace cuttlefish {
 namespace selector {
