+++ base/cvd/cuttlefish/host/commands/cvd/instances/BUILD.bazel
@@ -4,6 +4,26 @@ package(
     default_visibility = ["//:android_cuttlefish"],
 )
 
+cc_library(
+    name = "config_path",
+    srcs = ["config_path.cpp"],
+    hdrs = ["config_path.h"],
+    deps = [
+        "//cuttlefish/common/libs/utils:files",
+        "//cuttlefish/common/libs/utils:result",
+        "//libbase",
+    ],
+)
+
+clang_tidy_test(
+    name = "config_path_clang_tidy",
+    srcs = [":config_path"],
+    tags = [
+        "clang-tidy",
+        "clang_tidy",
+    ],
+)
+
 proto_library(
     name = "cvd_persistent_data_proto",
     srcs = ["cvd_persistent_data.proto"],
@@ -33,6 +53,25 @@ clang_tidy_test(
 )
 
 cc_library(
+    name = "device_name",
+    srcs = ["device_name.cpp"],
+    hdrs = ["device_name.h"],
+    deps = [
+        "//cuttlefish/common/libs/utils:result",
+        "//libbase",
+    ],
+)
+
+clang_tidy_test(
+    name = "device_name_clang_tidy",
+    srcs = [":device_name"],
+    tags = [
+        "clang-tidy",
+        "clang_tidy",
+    ],
+)
+
+cc_library(
     name = "instances",
     srcs = [
         "instance_database.cpp",
@@ -54,7 +93,6 @@ cc_library(
     ],
     deps = [
         ":cvd_persistent_data",
-        ":instance_database_utils",
         "//cuttlefish/common/libs/fs",
         "//cuttlefish/common/libs/utils:contains",
         "//cuttlefish/common/libs/utils:files",
@@ -70,7 +108,9 @@ cc_library(
         "//cuttlefish/host/commands/cvd/cli/commands:host_tool_target",
         "//cuttlefish/host/commands/cvd/cli/selector:parser",
         "//cuttlefish/host/commands/cvd/instances/lock",
+        "//cuttlefish/host/commands/cvd/instances:config_path",
         "//cuttlefish/host/commands/cvd/instances:data_viewer",
+        "//cuttlefish/host/commands/cvd/instances:device_name",
         "//cuttlefish/host/commands/cvd/instances:run_cvd_proc_collector",
         "//cuttlefish/host/commands/cvd/legacy:cvd_server_cc_proto",
         "//cuttlefish/host/commands/cvd/utils",
@@ -87,26 +127,6 @@ cc_library(
 clang_tidy_test(
     name = "instances_clang_tidy",
     srcs = [":instances"],
-    tags = [
-        "clang-tidy",
-        "clang_tidy",
-    ],
-)
-
-cc_library(
-    name = "instance_database_utils",
-    srcs = ["instance_database_utils.cpp"],
-    hdrs = ["instance_database_utils.h"],
-    deps = [
-        "//cuttlefish/common/libs/utils:files",
-        "//cuttlefish/common/libs/utils:result",
-        "//libbase",
-    ],
-)
-
-clang_tidy_test(
-    name = "instance_database_utils_clang_tidy",
-    srcs = [":instance_database_utils"],
     tags = [
         "clang-tidy",
         "clang_tidy",
