+++ base/cvd/cuttlefish/host/commands/cvd/instances/config_path.cpp
@@ -1,0 +1,35 @@
+/*
+ * Copyright (C) 2022 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#include "cuttlefish/host/commands/cvd/instances/config_path.h"
+
+#include <android-base/file.h>
+
+#include "cuttlefish/common/libs/utils/files.h"
+
+namespace cuttlefish {
+
+Result<std::string> GetCuttlefishConfigPath(const std::string& home) {
+  std::string home_realpath;
+  CF_EXPECT(DirectoryExists(home), "Invalid Home Directory");
+  CF_EXPECT(android::base::Realpath(home, &home_realpath));
+  static const char kSuffix[] = "/cuttlefish_assembly/cuttlefish_config.json";
+  std::string config_path = AbsolutePath(home_realpath + kSuffix);
+  CF_EXPECT(FileExists(config_path), "No config file exists");
+  return {config_path};
+}
+
+}  // namespace cuttlefish
